+++ zk_ip/guest/src/balance_tree.rs
